+++ CHANGELOG.md
@@ -11,11 +11,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- `GravityCollector.getSessionId()` provides the current session id.
+
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+- All code related to Cypress is gone in [gravity-cypress-plugin](https://github.com/Smartesting/gravity-cypress-plugin)
 
 ### Fixed